backend/requirements.txt
# NL2SQL自动化数据生成系统依赖包

# 数据库连接
pymysql>=1.1.0
psycopg2-binary>=2.9.0  # PostgreSQL支持（可选）
pyodbc>=4.0.0          # SQL Server支持（可选）

# LLM客户端
openai>=1.0.0

# SQL解析与验证
sqlglot>=20.0.0

# 配置文件
pyyaml>=6.0

# 工具库
requests>=2.31.0

#SQL解析工具
sqlparse
